move processors schema to own file
