+++ src/scripts/modules/configurations/utils/schemaValidation/validation.js
